# docs/swagger.yaml
basePath: /
definitions:
  models.Cart:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.CartItem:
    properties:
      cart_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.Chat:
    properties:
      created_at:
        type: string
      id:
        type: integer
      trainer_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Club:
    properties:
      address:
        type: string
      coordinates:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        description: Error message
        type: string
    type: object
...
  models.News:
    properties:
      address:
        type: string
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      news_date:
        type: string
      notify_users:
        type: boolean
      published_at:
        type: string
      start_time:
        type: string
      title:
        type: string
      updated_at:
        type: string
      vos_jdut:
        type: string
    type: object
...
  models.Order:
    properties:
      id:
        type: integer
      order_date:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
      what_paid_for:
        type: string
    type: object
  models.Payment:
    properties:
      amount:
        type: number
      id:
        type: integer
      order_id:
        type: integer
      payment_date:
        type: string
      payment_method:
        type: string
      user_id:
        type: integer
    type: object
...
host: 25.28.121.106:8080
info:
  contact: {}
  description: API для Basketball Academy
  title: Basketball Academy API
  version: "1.0"
paths:
...
  /api/Chat:
    get:
      consumes:
      - application/json
      description: Get a list of all Chat
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Chat'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all Chat
      tags:
      - Chat
    post:
      consumes:
      - application/json
      description: Create a new Chat with the given information
      parameters:
      - description: Chat data
        in: body
        name: Chat
        required: true
        schema:
          $ref: '#/definitions/models.Chat'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Chat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new Chat
      tags:
      - Chat
  /api/News:
    get:
      consumes:
      - application/json
      description: Get a list of all News
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.News'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all News
      tags:
      - News
    post:
      consumes:
      - application/json
      description: Create a new News with the given information
      parameters:
      - description: News data
        in: body
        name: News
        required: true
        schema:
          $ref: '#/definitions/models.News'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new News
      tags:
      - News
  /api/News/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a News by ID
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete News
      tags:
      - News
    put:
      consumes:
      - application/json
      description: Update News data by ID
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated News data
        in: body
        name: News
        required: true
        schema:
          $ref: '#/definitions/models.News'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update News
      tags:
      - News
...
  /api/cart-items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные товара в корзине
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.CartItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CartItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить товар в корзину
      tags:
      - Cart
  /api/carts/checkout:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID корзины
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Cart'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Оформить заказ из корзины
      tags:
      - Cart
  /api/carts/user/{userId}:
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cart'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить корзину пользователя по ID
      tags:
      - Cart
  /api/clubs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Club'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить все клубы
      tags:
      - Clubs
  /api/clubs/{id}:
    get:
      parameters:
      - description: ID клуба
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Club'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить клуб по ID
      tags:
      - Clubs
...
  /api/orders:
    get:
      consumes:
      - application/json
      description: Get a list of all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new Order with the given information
      parameters:
      - description: Order data
        in: body
        name: Order
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new Order
      tags:
      - orders
  /api/payments/{id}:
    get:
      parameters:
      - description: ID платежа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить детали платежа по ID
      tags:
      - Payments
  /api/payments/user/{userId}:
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Payment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить историю платежей пользователя
      tags:
      - Payments
...
swagger: "2.0"